+++ CHANGES.rst
@@ -12,6 +12,8 @@ Changelog
   [sgeulette]
 - Added edit icon on contact ajax cards.
   [chris-adam]
+- Added Signer/Approving icon on personnel listing table.
+  [chris-adam]
 
 3.1.1 (2026-03-27)
 ------------------